% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{particles_config}
\alias{particles_config}
\title{Config for particles.js}
\usage{
particles_config(
  particles.number.value = 80L,
  particles.number.density.enable = TRUE,
  particles.number.density.value_area = 800L,
  particles.color.value = "#112446",
  particles.shape.type = "circle",
  particles.shape.stroke.width = 0L,
  particles.shape.stroke.color = "#000000",
  particles.shape.polygon.nb_sides = 5L,
  particles.shape.image.src = "img/github.svg",
  particles.shape.image.width = 100L,
  particles.shape.image.height = 100L,
  particles.opacity.value = 0.5,
  particles.opacity.random = FALSE,
  particles.opacity.anim.enable = FALSE,
  particles.opacity.anim.speed = 1L,
  particles.opacity.anim.opacity_min = 0.1,
  particles.opacity.anim.sync = FALSE,
  particles.size.value = 3L,
  particles.size.random = TRUE,
  particles.size.anim.enable = FALSE,
  particles.size.anim.speed = 40L,
  particles.size.anim.size_min = 0.1,
  particles.size.anim.sync = FALSE,
  particles.line_linked.enable = TRUE,
  particles.line_linked.distance = 150L,
  particles.line_linked.color = "#112446",
  particles.line_linked.opacity = 0.4,
  particles.line_linked.width = 1L,
  particles.move.enable = TRUE,
  particles.move.speed = 6L,
  particles.move.direction = "none",
  particles.move.random = FALSE,
  particles.move.straight = FALSE,
  particles.move.out_mode = "out",
  particles.move.bounce = FALSE,
  particles.move.attract.enable = FALSE,
  particles.move.attract.rotateX = 600L,
  particles.move.attract.rotateY = 1200L,
  interactivity.detect_on = "canvas",
  interactivity.events.onhover.enable = TRUE,
  interactivity.events.onhover.mode = "repulse",
  interactivity.events.onclick.enable = TRUE,
  interactivity.events.onclick.mode = "push",
  interactivity.events.resize = TRUE,
  interactivity.modes.grab.distance = 400L,
  interactivity.modes.grab.line_linked.opacity = 1L,
  interactivity.modes.bubble.distance = 400L,
  interactivity.modes.bubble.size = 40L,
  interactivity.modes.bubble.duration = 2L,
  interactivity.modes.bubble.opacity = 8L,
  interactivity.modes.bubble.speed = 3L,
  interactivity.modes.repulse.distance = 200L,
  interactivity.modes.repulse.duration = 0.4,
  interactivity.modes.push.particles_nb = 4L,
  interactivity.modes.remove.particles_nb = 2L,
  retina_detect = TRUE
)
}
\arguments{
\item{particles.number.value}{Number of particles}

\item{particles.number.density.enable}{Logical, enable density}

\item{particles.number.density.value_area}{Density of particles.}

\item{particles.color.value}{Color of particles}

\item{particles.shape.type}{Type of shape(s): \code{"circle"}, \code{"edge"}, \code{"triangle"}, \code{"polygon"}, \code{"star"}, \code{"image"}}

\item{particles.shape.stroke.width}{Stroke width of shapes}

\item{particles.shape.stroke.color}{Stroke color of shapes}

\item{particles.shape.polygon.nb_sides}{For \code{"polygon"}, number of shapes}

\item{particles.shape.image.src}{For \code{"image"}, path to image}

\item{particles.shape.image.width}{For \code{"image"}, width of image}

\item{particles.shape.image.height}{For \code{"image"}, height of image}

\item{particles.opacity.value}{Opacity value}

\item{particles.opacity.random}{Logical. Random opacity}

\item{particles.opacity.anim.enable}{Enable animation for particles opacity}

\item{particles.opacity.anim.speed}{Speed of animation}

\item{particles.opacity.anim.opacity_min}{Minimal opacity}

\item{particles.opacity.anim.sync}{Synchronize opacity animation between particles}

\item{particles.size.value}{Size of particles}

\item{particles.size.random}{Logical. Ramdom size}

\item{particles.size.anim.enable}{Enable animation for size}

\item{particles.size.anim.speed}{Speed animation}

\item{particles.size.anim.size_min}{Minimal size when animated}

\item{particles.size.anim.sync}{Synchronize size animation between particles}

\item{particles.line_linked.enable}{Logical. Links particles to each other}

\item{particles.line_linked.distance}{Distance between particles}

\item{particles.line_linked.color}{Link color}

\item{particles.line_linked.opacity}{Link opacity}

\item{particles.line_linked.width}{Link width}

\item{particles.move.enable}{Logical.Should particles move ?}

\item{particles.move.speed}{Speed of particles}

\item{particles.move.direction}{Direction of particles: \code{"none"}, \code{"top"}, \code{"top-right"},
\code{"right"}, \code{"bottom-right"}, \code{"bottom"}, \code{"bottom-left"}, \code{"left"}, \code{"top-left"}}

\item{particles.move.random}{Logical radom movement}

\item{particles.move.straight}{Logical. Straight or not}

\item{particles.move.out_mode}{What does particles should do when reaching border of page ? \code{"out"} or \code{"bounce"}}

\item{particles.move.bounce}{Logical. Movement between particles}

\item{particles.move.attract.enable}{Logical}

\item{particles.move.attract.rotateX}{Numeric}

\item{particles.move.attract.rotateY}{Numeric}

\item{interactivity.detect_on}{On \code{"canvas"} or \code{"window"}}

\item{interactivity.events.onhover.enable}{Enable on hover with mouse}

\item{interactivity.events.onhover.mode}{What to do when mouse is hover: \code{"grab"}, \code{"bubble"}, \code{"repulse"}}

\item{interactivity.events.onclick.enable}{Enable on click with mouse}

\item{interactivity.events.onclick.mode}{What to do on click: \code{"push"} (add particles), \code{"remove"}, \code{"bubble"}, \code{"repulse"}}

\item{interactivity.events.resize}{Logical}

\item{interactivity.modes.grab.distance}{Grab particles within a distance}

\item{interactivity.modes.grab.line_linked.opacity}{Change opacity of links between particles grabbed}

\item{interactivity.modes.bubble.distance}{Transform particles to buble within a distance}

\item{interactivity.modes.bubble.size}{Size of bubles}

\item{interactivity.modes.bubble.duration}{Duration of effect}

\item{interactivity.modes.bubble.opacity}{Opacity of bubbles}

\item{interactivity.modes.bubble.speed}{Speed of bubbles}

\item{interactivity.modes.repulse.distance}{Distance to repulse particles}

\item{interactivity.modes.repulse.duration}{Duration of repulse}

\item{interactivity.modes.push.particles_nb}{Add how many new particles}

\item{interactivity.modes.remove.particles_nb}{Remove how many particles}

\item{retina_detect}{Logical}
}
\description{
See examples here \url{https://vincentgarreau.com/particles.js/}
}
\examples{

particles(particles_config(
  particles.color.value = "#DF0101",
  particles.shape.type = "star",
  particles.line_linked.color = "#DF0101"
))

}
