% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydemo.R
\name{shinydemo}
\alias{shinydemo}
\title{Launch Shiny demo}
\usage{
shinydemo(mode = c("container", "fullpage"))
}
\arguments{
\item{mode}{Example with \code{"fullpage"} particles and another one with particles inside a \code{"container"}.}
}
\description{
Launch Shiny demo
}
\examples{
if (interactive()) {

shinydemo("container")

shinydemo("fullpage")

}
}
