% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles.R
\name{particles}
\alias{particles}
\title{Particles background}
\usage{
particles(path_config = NULL)
}
\arguments{
\item{path_config}{Path to a config JSON for particles.js,
visit \url{https://vincentgarreau.com/particles.js/} to create one.}
}
\description{
Add a beautiful interactive particles background
 in Shiny apps and rmarkdown documents.
}
